Dodao kron i modale

+++ app/Http/Controllers/CityController.php
@@ -2,11 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use Exception;
+use App\Models\City;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\UpdateCityRequest;
 use App\Http\Requests\StoreCityRequest;
-use App\Models\City;
-use Exception;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class CityController extends Controller
 {
@@ -26,27 +27,52 @@ class CityController extends Controller
     // Čuvanje novog grada
     public function store(StoreCityRequest $request)
     {
-        City::create($request->Validated());
-
-
-        return redirect()->route('admin.cities.index')->with('success', 'Grad uspjesno napravljen');
+        try {
+            City::create($request->validated());
+            return redirect()->route('admin.cities.index')->with('modal', [
+            'message' => 'Grad uspjesno napravljen',
+            'type' => 'success'
+        ]);
+        } catch (Exception $e) {
+            return redirect()->back()->with('modal', [
+            'message' => 'Greška pri čuvanju grada',
+            'type' => 'error'
+            ]);
+        }
     }
 
 
     // Forma za editovanje postojećeg grada
     public function edit(string $id)
     {
-        $city = City::findOrFail($id);
-        return view('admin.cities.edit', compact('city'));
+        try {
+            $city = City::findOrFail($id);
+            return view('admin.cities.edit', compact('city'));
+        } catch (ModelNotFoundException $e) {
+            return redirect()->back()->with('modal', [
+            'message' => 'Greška pri čuvanju grada',
+            'type' => 'error'
+            ]);
+        }
     }
 
     // Update postojećeg grada
-    public function update(UpdateCityRequest $request, City $city)
+    public function update(UpdateCityRequest $request, string $id)
     {
+        try {
+            $city = City::findOrFail($id);
+            $city->update($request->validated());
 
-        $city->update($request->validated());
-
-        return redirect()->route('admin.cities.index')->with('success', 'Grad izmjenjen uspjesno');
+            return redirect()->route('admin.cities.index')->with('modal', [
+                'message' => 'Grad uspjesno izmjenjen',
+                'type' => 'success'
+            ]);
+        } catch (ModelNotFoundException $e) {
+            return redirect()->back()->with('modal', [
+            'message' => 'Greška pri čuvanju grada',
+            'type' => 'error'
+            ]);
+        }
     }
 
     // Brisanje grada
@@ -55,15 +81,24 @@ class CityController extends Controller
         try {
             $city = City::findOrFail($id);
 
-            if ($city->listings()->exists()) {
-                return redirect()->back()->with('error', 'Grad ima povezane oglase i ne može biti obrisan.');
+            if ($city->listings()->exists() || $city->reservations()->exists()) {
+                return redirect()->back()->with('modal', [
+            'message' => 'Greška pri čuvanju grada',
+            'type' => 'error'
+            ]);
             }
 
             $city->delete();
 
-            return redirect()->back()->with('success', 'Grad obrisan uspjesno');
+            return redirect()->back()->with('modal', [
+            'message' => 'Grad obrisan uspjesno',
+            'type' => 'success'
+            ]);
         } catch (Exception $e) {
-            return redirect()->back()->with('error', $e->getMessage());
+            return redirect()->back()->with('modal', [
+            'message' => 'Greška pri čuvanju grada',
+            'type' => 'error'
+            ]);
         }
     }
 }
